# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nutrition Clinic Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List all patients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientsListResponse"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientRequest"
      responses:
        "201":
          description: Patient created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{patientId}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient details
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Patient details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update patient details
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePatientRequest"
      responses:
        "200":
          description: Updated patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, sent, paid, void]
        - name: patientId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoicesListResponse"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create a new invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{invoiceId}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice details
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice status
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceRequest"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: appointmentId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: sendNotification
      tags: [notifications]
      summary: Send a notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendNotificationRequest"
      responses:
        "201":
          description: Notification sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /patient-files:
    post:
      operationId: createPatientFile
      tags: [files]
      summary: Create a patient file record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientFileRequest"
      responses:
        "201":
          description: File record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientFile"
  /patient-files/{patientId}:
    get:
      operationId: listPatientFiles
      tags: [files]
      summary: List patient files
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of patient files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PatientFile"
components:
  schemas:
...
    Patient:
      type: object
      properties:
        id:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        phone:
          type: string
        email:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female]
        nationalId:
          type: string
        medicalHistory:
          type: string
        allergies:
          type: string
        metadata:
          type: object
          description: Flexible extra fields (weight, height, measurements, etc.)
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        isActive:
          type: boolean
      required:
        - id
        - nameAr
        - phone
        - isActive
        - createdAt
    CreatePatientRequest:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        phone:
          type: string
        email:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female]
        nationalId:
          type: string
        medicalHistory:
          type: string
        allergies:
          type: string
        metadata:
          type: object
          additionalProperties: true
      required:
        - nameAr
        - phone
    UpdatePatientRequest:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        phone:
          type: string
        email:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female]
        nationalId:
          type: string
        medicalHistory:
          type: string
        allergies:
          type: string
        metadata:
          type: object
          additionalProperties: true
    PatientsListResponse:
      type: object
      properties:
        patients:
          type: array
          items:
            $ref: "#/components/schemas/Patient"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - patients
        - total
        - page
        - limit
...
    Invoice:
      type: object
      properties:
        id:
          type: string
        invoiceNumber:
          type: string
        patientId:
          type: string
        appointmentId:
          type: string
        subtotal:
          type: number
        vatAmount:
          type: number
        vatRate:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [draft, sent, paid, void]
        paymentMethod:
          type: string
          enum: [mada, apple_pay, stc_pay, cash, bank_transfer]
        issueDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        paidAt:
          type: string
          format: date-time
        notes:
          type: string
        patient:
          $ref: "#/components/schemas/Patient"
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - invoiceNumber
        - patientId
        - subtotal
        - vatAmount
        - vatRate
        - total
        - status
        - issueDate
        - createdAt
    InvoiceItem:
      type: object
      properties:
        id:
          type: string
        description:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        total:
          type: number
      required:
        - id
        - description
        - quantity
        - unitPrice
        - total
    CreateInvoiceRequest:
      type: object
      properties:
        patientId:
          type: string
        appointmentId:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              description:
                type: string
              quantity:
                type: integer
              unitPrice:
                type: number
            required:
              - description
              - quantity
              - unitPrice
        vatRate:
          type: number
          default: 0.15
        paymentMethod:
          type: string
          enum: [mada, apple_pay, stc_pay, cash, bank_transfer]
        dueDate:
          type: string
          format: date
        notes:
          type: string
      required:
        - patientId
        - items
    UpdateInvoiceRequest:
      type: object
      properties:
        status:
          type: string
          enum: [draft, sent, paid, void]
        paymentMethod:
          type: string
          enum: [mada, apple_pay, stc_pay, cash, bank_transfer]
        paidAt:
          type: string
          format: date-time
        notes:
          type: string
    InvoicesListResponse:
      type: object
      properties:
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - invoices
        - total
        - page
        - limit
...
    Notification:
      type: object
      properties:
        id:
          type: string
        appointmentId:
          type: string
        patientId:
          type: string
        type:
          type: string
          enum: [confirmation, reminder_24h, reminder_1h, follow_up, custom]
        channel:
          type: string
          enum: [whatsapp, sms, email]
        message:
          type: string
        status:
          type: string
          enum: [pending, sent, failed]
        sentAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - channel
        - status
        - createdAt
    SendNotificationRequest:
      type: object
      properties:
        appointmentId:
          type: string
        patientId:
          type: string
        type:
          type: string
          enum: [confirmation, reminder_24h, reminder_1h, follow_up, custom]
        channel:
          type: string
          enum: [whatsapp, sms, email]
        message:
          type: string
      required:
        - patientId
        - type
        - channel
    PatientFile:
      type: object
      properties:
        id:
          type: string
        patientId:
          type: string
        fileName:
          type: string
        fileType:
          type: string
          enum: [lab_result, nutrition_plan, prescription, other]
        fileUrl:
          type: string
        fileSize:
          type: integer
        uploadedAt:
          type: string
          format: date-time
        notes:
          type: string
      required:
        - id
        - patientId
        - fileName
        - fileType
        - uploadedAt
    CreatePatientFileRequest:
      type: object
      properties:
        patientId:
          type: string
        fileName:
          type: string
        fileType:
          type: string
          enum: [lab_result, nutrition_plan, prescription, other]
        fileUrl:
          type: string
        fileSize:
          type: integer
        notes:
          type: string
      required:
        - patientId
        - fileName
        - fileType