# lib/api-zod/src/generated/types/patientFile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nutrition Clinic Management API
 * OpenAPI spec version: 0.1.0
 */
import type { PatientFileFileType } from "./patientFileFileType";

export interface PatientFile {
  id: string;
  patientId: string;
  fileName: string;
  fileType: PatientFileFileType;
  fileUrl?: string;
  fileSize?: number;
  uploadedAt: Date;
  notes?: string;
}