# lib/api-zod/src/generated/types/updateInvoiceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nutrition Clinic Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateInvoiceRequestPaymentMethod } from "./updateInvoiceRequestPaymentMethod";
import type { UpdateInvoiceRequestStatus } from "./updateInvoiceRequestStatus";

export interface UpdateInvoiceRequest {
  status?: UpdateInvoiceRequestStatus;
  paymentMethod?: UpdateInvoiceRequestPaymentMethod;
  paidAt?: Date;
  notes?: string;
}